Add note about client type
We ran into an issue when using this project to perform auth for our chrome extension where we were getting 401 Unauthorized errors and couldn't figure out why. Turns out we'd incorrectly configured our client type. Wanted to put in a note so that others can avoid our silly rabbit hole in the future =).